backend/requirements.txt
# ───────────────────────────────
# Core Django
# ───────────────────────────────
Django~=5.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4
dj-database-url>=2.1

# ───────────────────────────────
# Authentication & Security
# ───────────────────────────────
djangorestframework-simplejwt>=5.3.1
django-allauth>=0.63
cryptography>=43.0
pyotp>=2.9.0
qrcode>=7.4
python-decouple>=3.8

# ───────────────────────────────
# Async & Background Tasks
# ───────────────────────────────
celery>=5.3.6,<6
redis>=5.0
django-celery-beat>=2.6,<3
django-celery-results>=2.5,<3

# ───────────────────────────────
# API Documentation
# ───────────────────────────────
drf-spectacular>=0.27

# ───────────────────────────────
# File Handling & Storage
# ───────────────────────────────
Pillow>=11.0
django-storages>=1.14
boto3>=1.34

# ───────────────────────────────
# Monitoring & Logging
# ───────────────────────────────
django-debug-toolbar>=4.3
sentry-sdk>=1.40

# ───────────────────────────────
# Payment Processing
# ───────────────────────────────
stripe>=10.0.0

# ───────────────────────────────
# Email & SMS
# ───────────────────────────────
django-anymail>=11.0
twilio>=9.0

# ───────────────────────────────
# Testing
# ───────────────────────────────
pytest>=8.0
pytest-django>=4.7
tzdata>=2025.2
vine>=5.1
